mm: kmemleak: Don't die when memory allocation fails
When memory is leaking, it's going to be harder to allocate more memory,
making it more likely for this failure condition inside of kmemleak to
manifest itself. This is extremely frustrating since kmemleak kills
itself upon the first instance of memory allocation failure.

Bypass that and make kmemleak more resilient when memory is running low.

Change-Id: Ief05e11bf5c687380eb8ccf4da8177708b875e38
Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>